refactor account funding in service main
do not use mut api reference to modify signer. create a new api for new signer.
